\name{colon}
\alias{colon}
\title{
Gene expression data for colon cancer
}
\description{
This data set consists of gene expression of colon cancer study. 
}
\usage{data(colon)}

\format{A matrix containing 2000 probe sets and 2 classes (T, F)}
\source{
Alon, U., Barkai, N., Notterman, D.A., Gish, K., Ybarra, S., Mack, D., Levine, A.J. (1999).
Broad Patterns of Gene Expression Revealed by Clustering Analysis of Tumor and Normal Colon Tissues 
probed by Oligonucleotide Arrays, PNAS, 96(12), 6745--6750.
}
\keyword{datasets}

