\name{leuk1}
\alias{leuk1}
\title{
Gene expression data for leukemia 
}
\description{
This data set consists of gene expression of leukemia study. 
}
\usage{data(leukemia)}

\format{A matrix containing 6817 probe sets and 38 samples (2 classes: AML, ALL)}
\source{
 Golub, T.R., Slonim, D.K., Tamayo, P., Huard, C., Gaasenbeek,
M., Mesirov, P., Coller, H., Loh, M.L., Downing, J.R., Caliguri,
M.A., Bloomfield, C.D., and Lander, E.S. (1999) Molecular
classification of cancer: class discovery and class prediction by
gene expression monitoring. Science, 286, 531-537.
}
\keyword{datasets}

